+++ warp/native/mesh.h
@@ -85,6 +85,12 @@ struct Mesh {
 
 CUDA_CALLABLE inline Mesh mesh_get(uint64_t id) { return *(Mesh*)(id); }
 
+CUDA_CALLABLE inline int mesh_get_group_root(uint64_t id, int group_id)
+{
+    Mesh* mesh = (Mesh*)(id);
+    return bvh_get_group_root((uint64_t)&mesh->bvh, group_id);
+}
+
 
 CUDA_CALLABLE inline Mesh& operator+=(Mesh& a, const Mesh& b)
 {
@@ -1488,14 +1494,15 @@ CUDA_CALLABLE inline bool mesh_query_ray(
     float& v,
     float& sign,
     vec3& normal,
-    int& face
+    int& face,
+    int root = -1
 )
 {
     Mesh mesh = mesh_get(id);
 
     int stack[BVH_QUERY_STACK_SIZE];
 
-    stack[0] = *mesh.bvh.root;
+    stack[0] = root == -1 ? *mesh.bvh.root : root;
     int count = 1;
 
     vec3 rcp_dir = vec3(1.0f / dir[0], 1.0f / dir[1], 1.0f / dir[2]);
@@ -1537,16 +1544,16 @@ CUDA_CALLABLE inline bool mesh_query_ray(
                     vec3 q = mesh.points[j];
                     vec3 r = mesh.points[k];
 
-                    float t, u, v, sign;
+                    float temp_t, temp_u, temp_v, temp_sign;
                     vec3 n;
 
-                    if (intersect_ray_tri_woop(start, dir, p, q, r, t, u, v, sign, &n)) {
-                        if (t < min_t && t >= 0.0f) {
-                            min_t = t;
+                    if (intersect_ray_tri_woop(start, dir, p, q, r, temp_t, temp_u, temp_v, temp_sign, &n)) {
+                        if (temp_t < min_t && temp_t >= 0.0f) {
+                            min_t = temp_t;
                             min_face = primitive_index;
-                            min_u = u;
-                            min_v = v;
-                            min_sign = sign;
+                            min_u = temp_u;
+                            min_v = temp_v;
+                            min_sign = temp_sign;
                             min_normal = n;
                         }
                     }
@@ -1573,6 +1580,68 @@ CUDA_CALLABLE inline bool mesh_query_ray(
     }
 }
 
+CUDA_CALLABLE inline bool
+mesh_query_ray_anyhit(uint64_t id, const vec3& start, const vec3& dir, float max_t, int root = -1)
+{
+    Mesh mesh = mesh_get(id);
+
+    int stack[BVH_QUERY_STACK_SIZE];
+
+    stack[0] = root == -1 ? *mesh.bvh.root : root;
+    int count = 1;
+
+    vec3 rcp_dir = vec3(1.0f / dir[0], 1.0f / dir[1], 1.0f / dir[2]);
+
+    const float eps = 1.e-3f;
+    float temp_t = 0.0f;
+    bool hit = false;
+
+    while (count) {
+        const int node_index = stack[--count];
+
+        BVHPackedNodeHalf lower = bvh_load_node(mesh.bvh.node_lowers, node_index);
+        BVHPackedNodeHalf upper = bvh_load_node(mesh.bvh.node_uppers, node_index);
+
+        // todo: switch to robust ray-aabb, or expand bounds in build stage
+        hit = intersect_ray_aabb(
+            start, rcp_dir, vec3(lower.x - eps, lower.y - eps, lower.z - eps),
+            vec3(upper.x + eps, upper.y + eps, upper.z + eps), temp_t
+        );
+
+        if (hit && temp_t < max_t) {
+            if (lower.b) {
+                const int start_index = lower.i;
+                const int end_index = upper.i;
+                // loops through primitives in the leaf
+                for (int primitive_counter = start_index; primitive_counter < end_index; primitive_counter++) {
+                    int primitive_index = mesh.bvh.primitive_indices[primitive_counter];
+                    int i = mesh.indices[primitive_index * 3 + 0];
+                    int j = mesh.indices[primitive_index * 3 + 1];
+                    int k = mesh.indices[primitive_index * 3 + 2];
+
+                    vec3 p = mesh.points[i];
+                    vec3 q = mesh.points[j];
+                    vec3 r = mesh.points[k];
+
+                    float temp_t, temp_u, temp_v, temp_sign;
+                    vec3 n;
+
+                    if (intersect_ray_tri_woop(start, dir, p, q, r, temp_t, temp_u, temp_v, temp_sign, &n)) {
+                        if (temp_t < max_t && temp_t >= 0.0f) {
+                            return true;
+                        }
+                    }
+                }
+            } else {
+                stack[count++] = lower.i;
+                stack[count++] = upper.i;
+            }
+        }
+    }
+
+    return false;
+}
+
 template <typename T> CUDA_CALLABLE inline void _swap(T& a, T& b)
 {
     T t = a;
@@ -1590,7 +1659,8 @@ CUDA_CALLABLE inline bool mesh_query_ray_ordered(
     float& v,
     float& sign,
     vec3& normal,
-    int& face
+    int& face,
+    int root = -1
 )
 {
     Mesh mesh = mesh_get(id);
@@ -1598,7 +1668,7 @@ CUDA_CALLABLE inline bool mesh_query_ray_ordered(
     int stack[BVH_QUERY_STACK_SIZE];
     float stack_dist[BVH_QUERY_STACK_SIZE];
 
-    stack[0] = *mesh.bvh.root;
+    stack[0] = root == -1 ? *mesh.bvh.root : root;
     stack_dist[0] = -FLT_MAX;
 
     int count = 1;
@@ -1615,13 +1685,13 @@ CUDA_CALLABLE inline bool mesh_query_ray_ordered(
     while (count) {
         count -= 1;
 
-        const int nodeIndex = stack[count];
-        const float nodeDist = stack_dist[count];
+        const int node_index = stack[count];
+        const float node_dist = stack_dist[count];
 
-        if (nodeDist < min_t) {
-            int left_index = mesh.bvh.node_lowers[nodeIndex].i;
-            int right_index = mesh.bvh.node_uppers[nodeIndex].i;
-            bool leaf = mesh.bvh.node_lowers[nodeIndex].b;
+        if (node_dist < min_t) {
+            int left_index = mesh.bvh.node_lowers[node_index].i;
+            int right_index = mesh.bvh.node_uppers[node_index].i;
+            bool leaf = mesh.bvh.node_lowers[node_index].b;
 
             if (leaf) {
                 const int start_index = left_index;
@@ -1637,16 +1707,16 @@ CUDA_CALLABLE inline bool mesh_query_ray_ordered(
                     vec3 q = mesh.points[j];
                     vec3 r = mesh.points[k];
 
-                    float t, u, v, w, sign;
+                    float temp_t, temp_u, temp_v, temp_sign;
                     vec3 n;
 
-                    if (intersect_ray_tri_rtcd(start, dir, p, q, r, t, u, v, w, sign, &n)) {
-                        if (t < min_t && t >= 0.0f) {
-                            min_t = t;
+                    if (intersect_ray_tri_woop(start, dir, p, q, r, temp_t, temp_u, temp_v, temp_sign, &n)) {
+                        if (temp_t < min_t && temp_t >= 0.0f) {
+                            min_t = temp_t;
                             min_face = primitive_index;
-                            min_u = u;
-                            min_v = v;
-                            min_sign = sign;
+                            min_u = temp_u;
+                            min_v = temp_v;
+                            min_sign = temp_sign;
                             min_normal = n;
                         }
                     }
@@ -1709,7 +1779,6 @@ CUDA_CALLABLE inline bool mesh_query_ray_ordered(
     }
 }
 
-
 CUDA_CALLABLE inline void adj_mesh_query_ray(
     uint64_t id,
     const vec3& start,
@@ -1721,6 +1790,7 @@ CUDA_CALLABLE inline void adj_mesh_query_ray(
     float sign,
     const vec3& n,
     int face,
+    int root,
     uint64_t adj_id,
     vec3& adj_start,
     vec3& adj_dir,
@@ -1731,6 +1801,7 @@ CUDA_CALLABLE inline void adj_mesh_query_ray(
     float& adj_sign,
     vec3& adj_n,
     int& adj_face,
+    int& adj_root,
     bool& adj_ret
 )
 {
@@ -1754,6 +1825,24 @@ CUDA_CALLABLE inline void adj_mesh_query_ray(
     );
 }
 
+CUDA_CALLABLE inline void adj_mesh_get_group_root(uint64_t id, int group_id, uint64_t&, int&, int&) { }
+
+CUDA_CALLABLE inline void adj_mesh_query_ray_anyhit(
+    uint64_t id,
+    const vec3& start,
+    const vec3& dir,
+    float max_t,
+    int root,
+    const bool& ret,
+    uint64_t adj_id,
+    vec3& adj_start,
+    vec3& adj_dir,
+    float& adj_max_t,
+    int& adj_root,
+    bool& adj_ret
+)
+{
+}
 
 // Stores the result of querying the closest point on a mesh.
 struct mesh_query_ray_t {
@@ -1790,11 +1879,12 @@ struct mesh_query_ray_t {
     bool result;
 };
 
-CUDA_CALLABLE inline mesh_query_ray_t mesh_query_ray(uint64_t id, const vec3& start, const vec3& dir, float max_t)
+CUDA_CALLABLE inline mesh_query_ray_t
+mesh_query_ray(uint64_t id, const vec3& start, const vec3& dir, float max_t, int root)
 {
     mesh_query_ray_t query;
     query.result
-        = mesh_query_ray(id, start, dir, max_t, query.t, query.u, query.v, query.sign, query.normal, query.face);
+        = mesh_query_ray(id, start, dir, max_t, query.t, query.u, query.v, query.sign, query.normal, query.face, root);
     return query;
 }
 
@@ -1803,20 +1893,21 @@ CUDA_CALLABLE inline void adj_mesh_query_ray(
     const vec3& start,
     const vec3& dir,
     float max_t,
+    int root,
     const mesh_query_ray_t& ret,
     uint64_t adj_id,
     vec3& adj_start,
     vec3& adj_dir,
     float& adj_max_t,
+    int& adj_root,
     mesh_query_ray_t& adj_ret
 )
 {
     adj_mesh_query_ray(
-        id, start, dir, max_t, ret.t, ret.u, ret.v, ret.sign, ret.normal, ret.face, adj_id, adj_start, adj_dir,
-        adj_max_t, adj_ret.t, adj_ret.u, adj_ret.v, adj_ret.sign, adj_ret.normal, adj_ret.face, adj_ret.result
+        id, start, dir, max_t, ret.t, ret.u, ret.v, ret.sign, ret.normal, ret.face, root, adj_id, adj_start, adj_dir,
+        adj_max_t, adj_ret.t, adj_ret.u, adj_ret.v, adj_ret.sign, adj_ret.normal, adj_ret.face, adj_root, adj_ret.result
     );
 }
-
 
 // determine if a point is inside (ret < 0 ) or outside the mesh (ret > 0)
 CUDA_CALLABLE inline float mesh_query_inside(uint64_t id, const vec3& p)
@@ -1828,7 +1919,9 @@ CUDA_CALLABLE inline float mesh_query_inside(uint64_t id, const vec3& p)
     int vote = 0;
 
     for (int i = 0; i < 3; ++i) {
-        if (mesh_query_ray(id, p, vec3(float(i == 0), float(i == 1), float(i == 2)), FLT_MAX, t, u, v, sign, n, face)
+        if (mesh_query_ray(
+                id, p, vec3(float(i == 0), float(i == 1), float(i == 2)), FLT_MAX, t, u, v, sign, n, face, -1
+            )
             && sign < 0) {
             vote++;
         }